Retrieve Self Assessment tax calculation income tax NICs calculated
Rebase

WIP

WIP

Rebase

Changes

fixed rebase error conflicts

fixes

Fixed json for get calc response

Added new test and renamed test

Added test JsError

added Json test error

changed the model values

changed big dec to int

changed student loans to array

+++ app/v1/models/response/getCalculation/incomeTaxAndNics/detail/GiftAid.scala
@@ -18,11 +18,10 @@ package v1.models.response.getCalculation.incomeTaxAndNics.detail
 
 import play.api.libs.json.{Json, OFormat}
 
-case class GiftAid(grossGiftAidPayments: BigDecimal,
+case class GiftAid(grossGiftAidPayments: BigInt,
                    rate: BigDecimal,
                    giftAidTax: BigDecimal)
 
 object GiftAid {
   implicit val format: OFormat[GiftAid] = Json.format[GiftAid]
 }
-
